@@ -29,24 +29,27 @@
     "@headlessui/react": "^1.6.6",
     "@hookform/resolvers": "^2.8.8",
     "@popperjs/core": "^2.11.5",
-    "@types/node": "^14.0.0",
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "13.4.0",
+    "@testing-library/user-event": "14.4.3",
+    "@types/node": "^16.0.0",
+    "@types/react": "18.0.27",
+    "@types/react-dom": "18.0.10",
     "ansi-to-react": "^6.1.6",
     "axios": "^1.1.3",
     "classnames": "^2.3.1",
     "immer": "^9.0.12",
     "lodash": "^4.17.21",
     "loglevel": "^1.8.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-hook-form": "^7.28.0",
     "react-hot-toast": "^2.2.0",
     "react-icons": "^4.3.1",
     "react-popper": "^2.2.5",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
-    "react-select": "^5.2.2",
+    "react-router-dom": "6.8.1",
+    "react-scripts": "5.0.1",
+    "react-select": "5.7.0",
     "reconnecting-websocket": "^4.4.0",
     "sort-keys": "^5.0.0",
     "use-persisted-state": "^0.3.3",
@@ -54,9 +57,6 @@
     "zod": "^3.20.2"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.3.0",
-    "@types/react-router-dom": "^5.3.3",
     "@types/use-persisted-state": "^0.3.1",
     "@typescript-eslint/eslint-plugin": "^5.9.0",
     "@typescript-eslint/parser": "^5.9.0",
@@ -72,8 +72,8 @@
     "eslint-plugin-react-hooks": "^4.3.0",
     "http-proxy-middleware": "^2.0.6",
     "prettier": "^2.5.1",
-    "prettier-plugin-tailwindcss": "^0.2.1",
-    "tailwindcss": "^3.0.23",
+    "prettier-plugin-tailwindcss": "0.2.1",
+    "tailwindcss": "3.2.4",
     "typescript": "4.5.4"
   },
   "engines": {
